// src/typing/enum/http-code.enum.ts
export enum HttpCodeEnum {
    Status200OK = 200,
    Status201Created = 201,
    Status204NotContent = 204,
    Status400BadRequest = 400,
    Status401Unauthorized = 401,
    Status403Forbidden = 403,
    Status404NotFound = 404,
    Status405MethodNotAllowed = 405,
    Status500ServerError = 500,
    Status504TimeOut = 504,
}